#2 If the server is not reachable, the HTTP status code 503 is sent to the client. If the server is not reachable within the timeout, the HTTP status code 504 is sent to the client.

+++ src/main/java/io/github/sebastianschmidt/galop/configuration/Configuration.java
@@ -12,7 +12,7 @@ public interface Configuration {
 
     PortNumber getTargetPort();
 
-    long getTargetConnectionTimeout();
+    int getTargetConnectionTimeout();
 
     int getMaxHttpHeaderSize();
