+++ src/main/java/io/github/sebastianschmidt/galop/configuration/ConfigurationImpl.java
@@ -11,7 +11,7 @@ final class ConfigurationImpl implements Configuration {
     private final PortNumber proxyPort;
     private final InetAddress targetAddress;
     private final PortNumber targetPort;
-    private long targetConnectionTimeout = ConfigurationDefaults.TARGET_CONNECTION_TIMEOUT;
+    private int targetConnectionTimeout = ConfigurationDefaults.TARGET_CONNECTION_TIMEOUT;
     private int maxHttpHeaderSize = ConfigurationDefaults.MAX_HTTP_HEADER_SIZE;
     private long connectionHandlersLogInterval = ConfigurationDefaults.CONNECTION_HANDLERS_LOG_INTERVAL;
     private long connectionHandlersTerminationTimeout = ConfigurationDefaults.CONNECTION_HANDLERS_TERMINATION_TIMEOUT;
@@ -38,11 +38,11 @@ final class ConfigurationImpl implements Configuration {
     }
 
     @Override
-    public long getTargetConnectionTimeout() {
+    public int getTargetConnectionTimeout() {
         return targetConnectionTimeout;
     }
 
-    void setTargetConnectionTimeout(final long targetConnectionTimeout) {
+    void setTargetConnectionTimeout(final int targetConnectionTimeout) {
         this.targetConnectionTimeout = targetConnectionTimeout;
     }
